1- Assigning relative scores to train stops
This rmd file is compiled of publicly available CTA data that can be found on the CTA's website. It uses a CTA Daily Train Ridership dataset, a CTA Train Frequency dataset, and a CTA Train Stop Coordinates Dataset. It uses this data to ultimately assign each train stop in the CTA transit a system a transit score that is based off a function of the stop's frequency and ridership numbers. The output of this code will yield a new csv file that will be used in step 2 (spatial analysis).

Note --> Daily train ridership dataset is too large of a file to attach. Must download the file from the first link directly below.

https://data.cityofchicago.org/Transportation/CTA-Ridership-L-Station-Entries-Daily-Totals/5neh-572f
https://www.transitchicago.com/developers/gtfs/
https://data.cityofchicago.org/Transportation/CTA-System-Information-List-of-L-Stops/8pix-ypme/about_data