1- Assigning relative scores to bus stops
This rmd file is compiled of publicly available CTA data that can be found on the CTA's website. It uses a CTA Bus Route Ridership dataset, a CTA Bus Frequency dataset, CTA Bus Stop Coordinates dataset, and a CTA Bus Stop dataset. It uses this data to ultimately assign each bus stop in the CTA transit a system a transit score that is based off a function of the stop's frequency and ridership numbers. The output of this code will yield a csv file that will be used for step 2 (spatial analysis).

Some limitations to mention is that we were not able to find ridership data for each bus stop, only by bus route. This led to us having to use the same daily data for all stops within a given route. In addition to this, the frequency data is for each stop on a single day, so it does not accurately take into account the day to day fluctuations in frequency numbers that are actually present.

https://data.cityofchicago.org/Transportation/CTA-Ridership-Bus-Routes-Daily-Totals-by-Route/jyb9-n7fm/about_data
https://www.transitchicago.com/developers/gtfs/
https://data.cityofchicago.org/Transportation/CTA-Bus-Stops/hvnx-qtky