2.1- Bus Spatial Analysis/ Residential Properties
The spatial analysis uses the bus monthly scores dataset that was created from step 1 for buses and the property values dataset. This files aim is to quantify transit accessibility for individual households in the Cook County region. This is done by plotting a the coordinates of a house, creating a 0.5 mile radius, counting the amount of bus stops captured in that radius, and then summing the relative scores of all Bus stops that were captured. This value that is calculated will be the property's transit score. This value will be the X variable for our analysis.

Note-- Due to computational limitations, we were not to assign every single property without taking samples and binding them together.
Note-- Also, the property values data set is too large to attach so please use link below to download data. The bus data relative scores dataset is also too big to attach here so 
            you must run step one for bus stops, download the csv file that is created at the end, and then use that file here.

https://datacatalog.cookcountyil.gov/Property-Taxation/Assessor-Archived-05-11-2022-Residential-Sales-Dat/5pge-nu6u