3- Rental Rates and 60602 Zipcode Analysis
These files aim at identifying the relationship between transit accessibility and rental rates. We first begin with the Annual Train Data.rmd file to create a dataset that we desire. We then move onto the Train Zipcode Spatial Join to assign each zipcode a quantified value determining its transit score. The remaining RMD files are all different analysis conducted. We found zipcode 60602 to be a zipcode of interest so many of thd files focus on this zipcode.
- CTA Train Ridership.csv is too large to attached. Please refer to link in previous section 2.1